pyproject.toml
[project]
name = "klim-cp-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg2>=2.9.11",
    "psycopg2-binary>=2.9.11",
    "pyqt6>=6.10.1",
    "python-docx>=1.2.0",
]
